Move string parsing to top level
Also create new class structure for tokenizer specs.
