+++ parser/parsedHedColumnSplice.js
