+++ parser/parsedHedString.js
