+++ tests/event.spec.js
@@ -2,8 +2,8 @@ import chai from 'chai'
 const assert = chai.assert
 import * as hed from '../validator/event'
 import { buildSchemas } from '../validator/schema/init'
-import { parseHedString } from '../validator/parser/main'
-import { ParsedHedTag } from '../validator/parser/parsedHedTag'
+import { parseHedString } from '../parser/main'
+import { ParsedHedTag } from '../parser/parsedHedTag'
 import { HedValidator, Hed2Validator, Hed3Validator } from '../validator/event'
 import { generateIssue } from '../common/issues/issues'
 import converterGenerateIssue from '../converter/issues'
