+++ validator/event/hed3.js
@@ -2,8 +2,8 @@ import differenceWith from 'lodash/differenceWith'
 import isEqual from 'lodash/isEqual'
 
 import { IssueError } from '../../common/issues/issues'
-import ParsedHedGroup from '../parser/parsedHedGroup'
-import { ParsedHedTag } from '../parser/parsedHedTag'
+import ParsedHedGroup from '../../parser/parsedHedGroup'
+import { ParsedHedTag } from '../../parser/parsedHedTag'
 import { getParsedParentTags } from '../../utils/hedData'
 import { getParentTag, getTagName, hedStringIsAGroup, replaceTagNameWithPound } from '../../utils/hedStrings'
 import { getCharacterCount, isNumber } from '../../utils/string'
